% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesisdown.R
\docType{package}
\name{huphydown}
\alias{huphydown}
\alias{huphydown-package}
\title{huphydown: A package for creating HU Physics undergraduate, Masters, and PhD theses using R Markdown}
\description{
Using the bookdown package and templates from
    HU Berlin Institute of Physics, this package aims to encourage reproducible research using
    simple Markdown syntax while embedding all of the R code to produce plots
    and analyses as well. Based on Chester Ismay's 
  `thesisdown` package.
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook
}

\author{
\strong{Maintainer}: Jakob Napiontek \email{jakob.napiontek@physik.hu-berlin.de}

}
