% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_word}
\alias{thesis_word}
\title{Creates an R Markdown Word Thesis document}
\usage{
thesis_word(...)
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}
}
\value{
A Word Document based on (hopefully soon, but not currently)
a HU Physics Word template
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a Microsoft Word version of the thesis.
}
\examples{
\dontrun{
output:huphydown::thesis_word
}
}
